+++ plugin/nube/protocals/bacnetmaster/bacnet.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"github.com/NubeDev/bacnet"
 	"github.com/NubeDev/bacnet/btypes"
 	"github.com/NubeDev/bacnet/btypes/segmentation"
@@ -9,6 +10,7 @@ import (
 	"github.com/NubeIO/flow-framework/utils/float"
 	"github.com/NubeIO/flow-framework/utils/integer"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
+	"github.com/iancoleman/strcase"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -90,7 +92,7 @@ func (inst *Instance) bacnetDevice(dev *model.Device) error {
 
 // getDev get an instance of a created bacnet device that is cached in bacnet lib
 func (inst *Instance) doReadValue(pnt *model.Point, networkUUID, deviceUUID string) (float64, error) {
-	object, _, isBool := setObjectType(pnt.ObjectType)
+	object, _, isBool, _ := setObjectType(pnt.ObjectType)
 	bp := &network.Point{
 		ObjectID:         btypes.ObjectInstance(integer.NonNil(pnt.ObjectId)),
 		ObjectType:       object,
@@ -133,7 +135,7 @@ func (inst *Instance) doReadValue(pnt *model.Point, networkUUID, deviceUUID stri
 
 func (inst *Instance) doWrite(pnt *model.Point, networkUUID, deviceUUID string) error {
 	val := float.NonNil(pnt.WriteValue)
-	object, isWrite, isBool := setObjectType(pnt.ObjectType)
+	object, isWrite, isBool, _ := setObjectType(pnt.ObjectType)
 	writePriority := integer.NonNil(pnt.WritePriority)
 	if writePriority == 0 {
 		writePriority = 16
@@ -176,29 +178,54 @@ func (inst *Instance) doWrite(pnt *model.Point, networkUUID, deviceUUID string) 
 
 }
 
-func setObjectType(object string) (obj btypes.ObjectType, isWritable, isBool bool) {
+func setObjectType(object string) (obj btypes.ObjectType, isWritable, isBool bool, asString string) {
+	object = strcase.ToSnake(object)
 	switch object {
 	case "analog_input":
-		return btypes.AnalogInput, false, false
+		return btypes.AnalogInput, false, false, "analog_input"
 	case "analog_output":
-		return btypes.AnalogOutput, true, false
+		return btypes.AnalogOutput, true, false, "analog_output"
 	case "analog_value":
-		return btypes.AnalogValue, true, false
+		return btypes.AnalogValue, true, false, "analog_value"
 	case "binary_input":
-		return btypes.BinaryInput, false, true
+		return btypes.BinaryInput, false, true, "binary_input"
 	case "binary_output":
-		return btypes.BinaryOutput, true, true
+		return btypes.BinaryOutput, true, true, "binary_output"
 	case "binary_value":
-		return btypes.BinaryValue, true, true
+		return btypes.BinaryValue, true, true, "binary_value"
 	case "multi_state_input":
-		return btypes.MultiStateInput, false, false
+		return btypes.MultiStateInput, false, false, "multi_state_input"
 	case "multi_state_output":
-		return btypes.MultiStateOutput, true, false
+		return btypes.MultiStateOutput, true, false, "multi_state_output"
 	case "multi_state_value":
-		return btypes.MultiStateValue, true, false
+		return btypes.MultiStateValue, true, false, "multi_state_value"
 	default:
-		return btypes.AnalogInput, false, false
+		return btypes.AnalogInput, false, false, "analog_input"
 	}
+}
+
+func convertObjectType(object btypes.ObjectType) string {
+	switch object {
+	case btypes.AnalogInput:
+		return "analog_input"
+	case btypes.AnalogOutput:
+		return "analog_output"
+	case btypes.AnalogValue:
+		return "analog_value"
+	case btypes.BinaryInput:
+		return "binary_input"
+	case btypes.BinaryOutput:
+		return "binary_output"
+	case btypes.BinaryValue:
+		return "binary_value"
+	case btypes.MultiStateInput:
+		return "multi_state_input"
+	case btypes.MultiStateOutput:
+		return "multi_state_output"
+	case btypes.MultiStateValue:
+		return "multi_state_value"
+	}
+	return "analog_input"
 }
 
 type SegmentedType string
@@ -225,6 +252,21 @@ func setSegmentation(SegmentedType string) (out segmentation.SegmentedType) {
 	}
 }
 
+func convertSegmentation(segmentedType segmentation.SegmentedType) SegmentedType {
+	switch segmentedType {
+	case segmentation.SegmentedBoth:
+		return SegmentedBoth
+	case segmentation.SegmentedTransmit:
+		return SegmentedTransmit
+	case segmentation.SegmentedReceive:
+		return SegmentedReceive
+	case segmentation.NoSegmentation:
+		return NoSegmentation
+	default:
+		return NoSegmentation
+	}
+}
+
 func (inst *Instance) doWriteBool(networkUUID, deviceUUID string, pnt *network.Point, value uint32) error {
 	net, err := inst.getBacnetNetwork(networkUUID)
 	if err != nil {
@@ -244,20 +286,41 @@ func (inst *Instance) doWriteBool(networkUUID, deviceUUID string, pnt *network.P
 
 }
 
-func (inst *Instance) whoIs(networkUUID string, opts *bacnet.WhoIsOpts) (resp []btypes.Device, err error) {
+func (inst *Instance) whoIs(networkUUID string, opts *bacnet.WhoIsOpts, addDevices bool) (resp []btypes.Device, err error) {
 	net, err := inst.getBacnetNetwork(networkUUID)
 	if err != nil {
 		return nil, err
 	}
-	go net.NetworkRun()
+	go net.NetworkRun() //TODO: do we need to defer a NetworkClose()?
+	defer net.NetworkClose()
 	devices, err := net.Whois(opts)
 	if err != nil {
 		return nil, err
 	}
+	if addDevices {
+		for _, device := range devices {
+			newDevice := &model.Device{
+				CommonName:     model.CommonName{Name: fmt.Sprintf("deviceId_%d_networkNum_%d", device.ID.Instance, device.NetworkNumber)},
+				DeviceMac:      integer.New(device.MacMSTP),
+				DeviceObjectId: integer.New(int(device.ID.Instance)),
+				NetworkNumber:  integer.New(device.NetworkNumber),
+				MaxADPU:        integer.New(int(device.MaxApdu)),
+				Segmentation:   string(convertSegmentation(segmentation.SegmentedType(device.Segmentation))),
+				NetworkUUID:    networkUUID,
+			}
+			addDevice, err := inst.addDevice(newDevice)
+			if err != nil {
+				log.Errorf("failed to add a new device from whois %d", device.ID.Instance)
+				return nil, err
+			}
+			log.Infof("added new device from whois %s", addDevice.Name)
+		}
+	}
+
 	return devices, err
 }
 
-func (inst *Instance) devicePoints(deviceUUID string) (resp []*network.PointDetails, err error) {
+func (inst *Instance) devicePoints(deviceUUID string, addPoints, writeablePoints bool) (resp []*network.PointDetails, err error) {
 	getNetwork, err := inst.db.GetNetworkByDeviceUUID(deviceUUID, api.Args{})
 	if err != nil {
 		return nil, err
@@ -266,7 +329,8 @@ func (inst *Instance) devicePoints(deviceUUID string) (resp []*network.PointDeta
 	if err != nil {
 		return nil, err
 	}
-	go net.NetworkRun()
+	go net.NetworkRun() //TODO: do we need to defer a NetworkClose()?
+	defer net.NetworkClose()
 	dev, err := inst.getBacnetDevice(deviceUUID)
 	if err != nil {
 		return nil, err
@@ -276,8 +340,29 @@ func (inst *Instance) devicePoints(deviceUUID string) (resp []*network.PointDeta
 	if err != nil {
 		return nil, err
 	}
+	if addPoints {
+		for _, pnt := range resp {
+			_, isWrite, _, objectType := setObjectType(convertObjectType(pnt.ObjectType))
+			writeMode := model.ReadOnly
+			if isWrite && writeablePoints {
+				writeMode = model.WriteOnceThenRead
+			}
+			newPnt := &model.Point{
+				CommonName: model.CommonName{Name: pnt.Name},
+				DeviceUUID: deviceUUID,
+				ObjectType: objectType,
+				ObjectId:   integer.New(int(pnt.ObjectID)),
+				WriteMode:  writeMode,
+			}
+			point, err := inst.addPoint(newPnt)
+			if err != nil {
+				log.Errorf("failed to add a new point from discover points %s", pnt.Name)
+				return nil, err
+			}
+			log.Infof("added new point from discover points%s", point.Name)
+		}
+	}
 	return
-
 }
 
 func intToUnit32(value int) uint32 {