+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
+## [v0.5.5](https://github.com/NubeIO/flow-framework/tree/v0.5.5) (2022-06-10)
+
+- redo of bacnet-server plugin
+- fix bug on lora plugin
+- reformat nil helpers
+- add in sync on flow-networks
 
 ## [v0.5.6](https://github.com/NubeIO/flow-framework/tree/v0.5.6) (2022-06-17)